fix: removed tut_utils, shortened subjects fields and changed clip ex… (#374)
* fix: removed tut_utils, shortened subjects fields and changed clip extractions logic

To improve the architecture of kso I moved all the functions out of tutorials_utils as broken down below. The reason is we shouldn't have "tutorial-specific" functions but rather functions organised by key elements of the system (e.g. yolo, zooniverse, movies, frames, widgets...) that can be called from the projectprocessor. I also removed functions we no longer use.

Functions removed:
- write_urls_to_file
- get_project_info
- is_url (now we use the function from ultralytics)
- expand_list (now we use the inbuilt explode func)
- check_clip_size
- get_species_mapping
- choose_model_type

Functions moved to:
Widgets:
- log_meta_changes
- view_frames
- compare_frames
- view_frames
- launch_viewer (now launch_classifications_viewer)
- view_subject
- explore_classifications_per_subject
- choose_eval_params
- choose_test_prop
- choose_train_params
- choose_experiment_name
- choose_entity
- choose_conf
- choose_text
- WidgetMaker
- format_to_gbif

Zoo:
- check_clip_size
- expand_list
- create_clips
- extract_clips
- check_frame_size
- launch_table (now launch_classifications_table)

Movie:
- check_ffmpeg_availability

Frame:
- draw_annotations_in_frame

Yolo:
- encode_image
- get_annotations_viewer
- get_species_mapping
- process_detections
- plot_processed_detections

Removed workflow_id, classifications_count, retired_at and retirement_reason from the subjects table in schema because we now retrieve this information from the actual classifications rather than the subject information stored in Zooniverse. The subject information stored in Zooniverse is not accurate as we have modified workflows' retirement criteria multiple times and removed some subjects or used different format.

I cleaned the process to create clips to upload to Zooniverse by removing the redundant sample_subjects_from_workflows function

I broke down the interactive choose_footage to three steps to facilitate debugging and maintenance. Now users need to 1) choose whether they are using existing or new footage, 2) select the footage of interest and 3) there is a function (check_selected_movies) to store in the project_processor the paths, names and ids of the movies selected.

I tidied imports and text format using flake8

* fixed:removed tutorial_utils

* added str approach from commit 6fdb3ae
