+++ kso_utils/db_starter/schema.py
@@ -33,7 +33,7 @@ PhotoPosition int NULL,
 siteName text NULL,
 SurveyID int NULL,
 UNIQUE (filename)
-); 
+);
 
 CREATE TABLE IF NOT EXISTS subjects
 (
@@ -44,11 +44,7 @@ clip_start_time datetime NULL,
 clip_end_time datetime NULL,
 frame_exp_sp_id integer NULL,
 frame_number integer NULL,
-workflow_id varchar(255) NULL,
 subject_set_id varchar(255),
-classifications_count integer NULL,
-retired_at datetime NULL,
-retirement_reason text NULL,
 created_at datetime,
 https_location text NULL,
 movie_id integer NULL,
