+++ kso_utils/db_utils.py
@@ -1,7 +1,7 @@
 # base imports
-from pathlib import Path
 import sqlite3
 import logging
+from pathlib import Path
 import pandas as pd
 
 # util imports
@@ -394,8 +394,7 @@ def process_test_csv(
     # Rename id columns using the dictionary
     local_df = local_df.rename(columns=id_lookup)
 
-    ##################
-    ##Roadblock to ensure cols match schema
+    # Roadblock to ensure cols match schema
     ##################
     # Get the "standard" schema column names of the table of interest
     col_names_dic = get_column_names_db(conn, init_key)
@@ -534,12 +533,9 @@ def add_db_info_to_df(
 
     # If there are missing values, raise an issue
     if missing_values.any():
-        # Get the indices of the missing values
-        missing_indices = missing_values[missing_values].index
-
-        # Print a warning or raise an exception with relevant information
+        # Log a warning or raise an exception with relevant information
         logging.error(
-            f"Some rows in df do not have corresponding values in sql_df. Rows with missing values: {missing_indices}"
+            f"Some rows in df do not have corresponding values in sql_df. Rows with missing values are: {comb_df[missing_values]}"
         )
 
     # Drop the id column to prevent duplicated column issues
