+++ kso_utils/server_utils.py
@@ -1,12 +1,10 @@
 # base imports
 import os
-import pandas as pd
 import getpass
 import gdown
 import zipfile
 import boto3
 import logging
-import sys
 from tqdm import tqdm
 from pathlib import Path
 
@@ -122,14 +120,14 @@ def download_init_csv(project: Project, init_keys: list, server_connection: dict
             db_initial_info[str("server_" + init_key + "_csv")] = server_csv
 
             # Specify the local path for the csv
-            local_i_csv = str(Path(project.csv_folder, Path(server_csv).name))
+            local_i_csv = Path(project.csv_folder, Path(server_csv).name)
 
             # Download the csv
             download_object_from_s3(
                 client=server_connection["client"],
                 bucket=project.bucket,
                 key=server_csv,
-                filename=local_i_csv,
+                filename=str(local_i_csv),
             )
 
     else:
@@ -394,6 +392,7 @@ def download_object_from_s3(
             Key=key,
             ExtraArgs=ExtraArgs,
             Filename=filename,
+            Config=boto3.s3.transfer.TransferConfig(use_threads=False),
             Callback=lambda bytes_transferred: pbar.update(bytes_transferred),
         )
 
@@ -450,9 +449,9 @@ def delete_file_from_s3(client: boto3.client, *, bucket: str, key: str):
     client.delete_object(Bucket=bucket, Key=key)
 
 
-###################################
-########Google Drive functions#####
-###################################
+#####################
+# ## Google Drive functions ###
+# ####################
 
 
 def download_gdrive(gdrive_id: str, folder_name: str, fix_encoding: bool = True):
