+++ test/notebook-tests.py
@@ -1,24 +1,23 @@
 # -*- coding: utf-8 -*-
 """
 
-This auto-test does not test if everything displays what it should display. 
+This auto-test does not test if everything displays what it should display.
 It mainly tests if everything still runs without giving any errors.
 
 This collection of tests covers the following tutorial notebooks:
  - 1, 3, 4, 5, 6, 8
 
 All other tutorials are not tested automatically and care should be taken when
-making changes as these could break existing workflows. 
+making changes as these could break existing workflows.
 
 To run these tests manually, use pytest --disable-warnings test/notebook-tests.py
 
 """
 
 # --------------------The first 2 cells of every notebook-----------------------
-## Import Python packages
+# Import Python packages
 import os
 import sys
-import subprocess
 from pathlib import Path
 
 # Set environment variables
@@ -109,16 +108,16 @@ def test_t3(zoo_user, zoo_pass):
     # Connect to zooniverse w/ Github credential
     pp.connect_zoo_project(zoo_cred=[zoo_user, zoo_pass])
     # Pre-selected test movie
-    pp.movies_selected = ["movie_1.mp4"]
-    pp.movies_paths = ["https://www.wildlife.ai/wp-content/uploads/2022/06/movie_1.mp4"]
+    pp.selected_movies = ["movie_1.mp4"]
+    pp.selected_movies_paths = [
+        "https://www.wildlife.ai/wp-content/uploads/2022/06/movie_1.mp4"
+    ]
     # Check whether movie has been uploaded previously
-    pp.check_movies_uploaded(pp.movies_selected)
+    pp.check_movies_uploaded_zoo()
     # Do not use GPU by default
     gpu_available = kso_widgets.gpu_select()
     # Generate a default number of clips for testing
     pp.generate_zoo_clips(
-        movies_selected=pp.movies_selected,
-        movies_paths=pp.movies_paths,
         is_example=True,
         use_gpu=gpu_available.result,
         test=True,
@@ -133,8 +132,6 @@ def test_t3(zoo_user, zoo_pass):
 # since we do not want to upload things to them all the time.
 # """
 def test_t4(zoo_user, zoo_pass):
-    import kso_utils.tutorials_utils as t_utils
-
     # import kso_utils.widgets as kso_widgets
 
     # Log into Zooniverse and load relevant data from their DB
@@ -146,7 +143,7 @@ def test_t4(zoo_user, zoo_pass):
     # Fetch relevant frame subjects in dataframe (by default all species for testing)
     pp.extract_zoo_frames(test=True)
     # Review the size of the clips
-    t_utils.check_frame_size(frame_paths=pp.generated_frames["frame_path"].unique())
+    pp.check_frame_size()
     # Generate frames from based on subject metadata
     pp.modify_zoo_frames(test=True)
     # input_folder = kso_widgets.choose_folder()
@@ -161,11 +158,9 @@ def test_t4(zoo_user, zoo_pass):
     #    frames_skip=None,
     # )
     # Ensure that extracted and modified frames are of suitable size for upload to ZU
-    t_utils.check_frame_size(
-        frame_paths=pp.modified_frames["modif_frame_path"].unique()
-    )
+    pp.check_frame_size()
     # Compare original vs modified frames (no interactivity tested)
-    t_utils.compare_frames(df=pp.modified_frames)
+    pp.compare_frames(df=pp.modified_frames)
     # Test that final generated frames contain 9 rows (representing a classification)
     assert len(pp.modified_frames) == 9
 
@@ -175,12 +170,10 @@ def test_t4(zoo_user, zoo_pass):
 # ...
 # """
 def test_t5():
-    import kso_utils.tutorials_utils as t_utils
-    import kso_utils.yolo_utils as y_utils
     import kso_utils.server_utils as s_utils
+    import kso_utils.widgets as kso_widgets
 
     # Generate current timestamp
-    import shutil
     from datetime import datetime
 
     dt = datetime.now()
@@ -212,12 +205,12 @@ def test_t5():
     assert len(list(Path(exp_path, "weights").glob("*"))) == 2
 
     # Model evaluation
-    conf_thres = t_utils.choose_eval_params()
+    conf_thres = kso_widgets.choose_eval_params()
     # Evaluate YOLO Model on Unseen Test data
     mlp.eval_yolo(exp_name=exp_name, conf_thres=conf_thres.value)
 
     # Enhancement tests (leave out for now)
-    # eh_conf_thres = t_utils.choose_eval_params()
+    # eh_conf_thres = widgets.choose_eval_params()
     # mlp.enhance_yolov5(conf_thres=eh_conf_thres.value,
     #                   in_path=os.path.join(mlp.output_path, "ml-template-data"),
     #                   project_path=project_path, img_size=[640, 640])
@@ -236,11 +229,10 @@ def test_t5():
 
 
 def test_t6():
-    import kso_utils.tutorials_utils as t_utils
     import kso_utils.server_utils as s_utils
+    import kso_utils.widgets as kso_widgets
 
     # Generate current timestamp
-    import shutil
     from datetime import datetime
 
     dt = datetime.now()
@@ -248,7 +240,6 @@ def test_t6():
     # Create a unique experiment name
     exp_name = f"custom_{dt}".replace(" ", "_").replace(".", "_").replace(":", "-")
     project_path = str(Path(mlp.output_path, mlp.project_name))
-    exp_path = str(Path(project_path, exp_name))
 
     # Evaluation
     s_utils.get_ml_data(project, test=True)
@@ -257,7 +248,7 @@ def test_t6():
     artifact_dir = mlp.get_model(model, mlp.output_path)
     source = str(Path("../test/test_output", mlp.project.ml_folder, "images"))
     save_dir = project_path
-    conf_thres = t_utils.choose_conf()
+    conf_thres = kso_widgets.choose_conf()
     mlp.detect_yolo(
         project=project_path,
         name=exp_name,
@@ -293,7 +284,7 @@ def test_t6():
 # #-------------Tutorial 7-------------------------------------------------------
 """
 This tutorial is still in a very early stage and mostly uses widgets directly from widgets.py and
-API functionality from Zenodo. Tests should be added in future as its capabilities expand. 
+API functionality from Zenodo. Tests should be added in future as its capabilities expand.
 """
 
 
@@ -330,27 +321,9 @@ def test_t8(zoo_user, zoo_pass):
     # from workflows
     assert len(class_df) == 9
 
-    import kso_utils.zooniverse_utils as zoo_utils
-
     # Retrieve a subset of the subjects from the workflows of interest and
     # populate the sql subjects table
-    selected_zoo_workflows = zoo_utils.sample_subjects_from_workflows(
-        project=pp.project,
-        server_connection=pp.server_connection,
-        db_connection=pp.db_connection,
-        workflow_widget_checks=workflow_checks,
-        workflows_df=pp.zoo_info["workflows"],
-        subjects_df=pp.zoo_info["subjects"],
-    )
-
-    zoo_utils.process_zoo_classifications(
-        project=pp.project,
-        db_connection=pp.db_connection,
-        csv_paths=pp.csv_paths,
-        classifications_data=pp.zoo_info["classifications"],
-        subject_type=workflow_checks["Subject type: #0"],
-        selected_zoo_workflows=selected_zoo_workflows,
-    )
+    pp.process_zoo_classifications(test=True)
 
     agg_params = kso_widgets.choose_agg_parameters(workflow_checks["Subject type: #0"])
